完成微信支付对接

+++ aspnet-core/src/JustERP.Core.User/Payments/ExpertOrderPaymentManager.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Abp.Domain.Repositories;
 using Abp.Domain.Services;
 using JustERP.Core.User.Orders;
@@ -18,6 +19,11 @@ namespace JustERP.Core.User.Payments
 
         public async Task<LhzxExpertOrderPayment> CreateOrder(LhzxExpertOrderPayment orderPayment, LhzxExpertOrder order)
         {
+            var existsPayment = await _orderPaymentRepository.FirstOrDefaultAsync(p => p.ExpertOrderId == order.Id);
+            if (existsPayment != null && existsPayment.Status == (int)PaymentStatus.PayComplete)
+            {
+                throw new ApplicationException("已支付的订单无法再重新发起支付");
+            }
             await _orderPaymentRepository.DeleteAsync(p => p.ExpertOrderId == order.Id);
 
             orderPayment.Create(order);
@@ -25,6 +31,17 @@ namespace JustERP.Core.User.Payments
             await _orderPaymentRepository.InsertAsync(orderPayment);
 
             return orderPayment;
+        }
+
+        /// <summary>
+        /// 支付成功
+        /// </summary>
+        /// <param name="orderPayment"></param>
+        /// <returns></returns>
+        public async Task PaymentComplete(LhzxExpertOrderPayment orderPayment)
+        {
+            orderPayment.Status = (short)PaymentStatus.PayComplete;
+            await _orderPaymentRepository.UpdateAsync(orderPayment);
         }
     }
 }
